requirements.txt
numpy
pandas
matplotlib
nltk
pyspellchecker
gensim
scikit-learn
scipy
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
idna==2.8
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
PyYAML==5.3
requests==2.22.0
ruamel-yaml==0.15.46
six==1.12.0
tqdm==4.36.1
urllib3==1.24.2
pyphen
